+++ include/audio.hh
@@ -1,9 +1,9 @@
-#ifndef AUDIO_AUDIO_HH
-#define AUDIO_AUDIO_HH
+#ifndef AUDIO_HH
+#define AUDIO_HH
 
 #include <cstdint>
 #include <vector>
 
 using Audio = std::vector<int16_t>;
 
-#endif // AUDIO_AUDIO_HH
+#endif // AUDIO_HH
