+++ src/filter.cc
@@ -1,4 +1,4 @@
-#include "../include/filter/filter.hh"
+#include "../include/filter.hh"
 
 BPFilter::BPFilter(int order, double sampling_rate, double centre_freq_, double freq_range_, Pipe<Audio> &input_,
                    Pipe<Audio> &output_, std::chrono::milliseconds timeout_)
